requirement.txt
asgiref==3.2.3
Django==3.0
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
Pillow==6.2.1
PyJWT==1.7.1
pytz==2019.3
sqlparse==0.3.0
